# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "National Exam Card API",
        "description": "The National Exam Card API provides access to student's application for exam card data and process. ",
        "version": "0.1.0"
    },
    "paths": {
...
        "/paper": {
            "get": {
                "tags": [
                    "paper"
                ],
                "summary": "Get a list of all papers",
                "operationId": "d58075c7926292a5de2a63061bde3e55",
                "responses": {
                    "200": {
                        "description": "List of papers"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "paper"
                ],
                "summary": "Create a new paper",
                "operationId": "3941b36bac78c8f5f9cf4d07806e3ca3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Sample Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Sample content"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Past paper created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/paper/{id}": {
            "get": {
                "tags": [
                    "paper"
                ],
                "summary": "Get a specific paper by ID",
                "operationId": "bae445e170aa0aa12a0965d1ad868924",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the paper",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Past paper details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "paper"
                ],
                "summary": "Update an existing past paper",
                "operationId": "c032fe60d70805ccd1fc5a25e6af8e81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the past paper",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated content"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Past paper updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "paper"
                ],
                "summary": "Delete a past paper by ID",
                "operationId": "da342e2727072bc7f428d8dd23d5ca0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the past paper",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Past paper deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}